initializing multi lrt functions
we need to search efficiently through the range of
cut-points available in the linear predictions from
fitted coxph models. It is not clear whether we should
just recompute the LRT or set aside a matrix in memory
that would make recomputation less intense (but require
more memory). We need a benchmark to compare